requirements.txt
pydicom==2.4.1
Pillow==9.2.0
reportlab==3.6.12
numpy==1.21.2